refactor: move logic to UserService and remove unused dependencies
