+++ backend/src/main/java/com/happyplants/Application.java
@@ -1,7 +1,6 @@
 package com.happyplants;
 
-import io.github.cdimascio.dotenv
-        .Dotenv;
+import io.github.cdimascio.dotenv.Dotenv;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
