+++ frontend/src/pages/(auth-protected)/profile/+Page.tsx
@@ -1,5 +1,5 @@
-import { Button } from "@/components/ui/button";
-import { toast } from "sonner";
+import {Button} from "@/components/ui/button";
+import {toast} from "sonner";
 import {useEffect, useState} from "react";
