+++ frontend/src/pages/(auth-protected)/profile/change-password/+Page.tsx
@@ -1,7 +1,7 @@
-import { useState, type SyntheticEvent } from 'react';
-import { toast } from 'sonner';
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
+import {type SyntheticEvent, useState} from 'react';
+import {toast} from 'sonner';
+import {Button} from "@/components/ui/button";
+import {Input} from "@/components/ui/input";
 
 export default function ChangePasswordPage() {
     const [currentPassword, setCurrentPassword] = useState('');
