+++ frontend/src/pages/+Layout.tsx
@@ -2,7 +2,7 @@ export { Layout };
 
 import React from "react";
 import "@/css/shadcn.css";
-import { Toaster } from "@/components/ui/sonner";
+import {Toaster} from "@/components/ui/sonner";
 
 
 function Layout({ children }: { children: React.ReactNode }) {
