+++ frontend/src/pages/auth/register/+Page.tsx
@@ -1,12 +1,12 @@
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Leaf } from "lucide-react";
-import { navigate } from "vike/client/router";
-import { API_BASE_URL } from "@/config";
-import { toast } from "sonner";
+import {useState} from "react";
+import {Button} from "@/components/ui/button";
+import {Input} from "@/components/ui/input";
+import {Label} from "@/components/ui/label";
+import {Card, CardContent, CardHeader, CardTitle} from "@/components/ui/card";
+import {Leaf} from "lucide-react";
+import {navigate} from "vike/client/router";
+import {API_BASE_URL} from "@/config";
+import {toast} from "sonner";
 
 export default function Page() {
   const [email, setEmail] = useState("");
